requirements.txt
fastapi==0.111.0
fastapi-cli==0.0.4
python-dotenv==1.0.1
PyMySQL==1.1.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
loguru==0.7.2
uvicorn==0.24.0.post1
